% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_points_subset.R
\name{download_points_subset}
\alias{download_points_subset}
\title{Request dataset polygon subset}
\usage{
download_points_subset(
  dsvars,
  dates = NULL,
  years = NULL,
  months = NULL,
  days = NULL,
  t_crs = NULL,
  t_geom = NULL,
  out_format = "csv",
  grain_method = "strict",
  validate_method = "strict",
  ri_method = "nearest",
  dsn = NULL,
  req_name = NULL
)
}
\arguments{
\item{dsvars}{The GeoCDL dataset and variable IDs as a data.frame or matrix
with the first column specifying the datset ID and the second column
specifying the variable IDs. One row per unique dataset-variable,
combination.}

\item{dates}{The dates for which to request data. Dates must be specified as
strings, where "YYYY" means extract annual data, "YYYY-MM" is for monthly
data, and "YYYY-MM-DD" is for daily data. Date ranges can be specified using
a colon, as in "YYYY:YYYY", "YYYY-MM:YYYY-MM", or "YYYY-MM-DD:YYYY-MM-DD".
Leading 0s in "MM" and "DD" are optional (e.g., both "2020-01-01" and
"2020-1-1" are valid). Multiple dates and/or date ranges should be separated
by commas. For example: "2000:2010", "2000-1,2005-1,2010-1:2010-6",
"2000-01-01:2000-04-31". Dates can be omitted for non-temporal data
requests. For more complicated, recurring date patterns, use the "years",
"months", and "days" parameters instead of `dates`. See the Details section
below for more information about temporal subsetting.}

\item{years}{Years to subset data. Can be expressed as ranges and sequences,
such as "2004:2005,2009" or "2000:2010+2", which is interpreted as every
other year starting with 2000. Ranges are inclusive of their endpoints
unless the endpoint does not correspond with the step size increment. If
"dates" is also provided, "years" (and "months" and "days") will be ignored.
See the Details section below for more information about temporal subsetting.}

\item{months}{Months to include for each year of the data request. Only
valid if "years" is also specified. Accepts values 1-12. Can be expressed
as ranges and sequences, such as "1:3,5,9:12" or "1:12+2", which is
interpreted as every other month. Ranges are inclusive of their endpoints
unless the endpoint does not correspond with the step size increment.
See the Details section below for more information about temporal subsetting.}

\item{days}{Only valid if "years" or "years" and "months" are also specified.
If only "years" is specified, "days" is interpreted as the days of each
year (starting from 1) to include in the request. If "years" and "months"
are both specified, "days" is interpreted as the days of each month
(starting from 1) to include in the request. The special value "N"
represents the last day of a month or year. Can be expressed as ranges and
sequences, such as "1:100,200:230,266:366", "1:N", or "10:N+10", which is
interpreted as every 10th day of the year or month. Ranges are inclusive
of their endpoints unless the endpoint does not correspond with the step
size increment. Required if "years" or "years" and "months" are specified
and daily data are desired. See the Details section below for more
information about temporal subsetting.}

\item{t_crs}{The target coordinate reference system (CRS) for the returned
data. Can be specified as a PROJ string, CRS WKT string, authority string
(e.g., "EPSG:4269"), or PROJ object name (e.g., "NAD83").}

\item{t_geom}{The point geometry of interest specified as a geometry upload
ID (see \code{\link{upload_geometry}}), `sf` or `sp` object, or a data.frame
with x and y columns. If a data.frame, the CRS of the points are assumed to
be that of `t_crs`, if specified, or that of the first requested dataset.}

\item{out_format}{The output format for the subset: 'csv', 'shapefile', or
'netcdf'.}

\item{grain_method}{How to handle scenario of requested date grains not
matching date grains of each requested dataset. If "strict" (default),
an error will be returned. If "skip", the dataset will be skipped.
If "coarser", then only coarser grains will be returned. If "finer",
then only finer grains will be returned. If "any", then any available
grain will be returned, with coarser having higher priority over finer.
Non-temporal datasets are always returned. See the Details section below
for more information about temporal subsetting.}

\item{validate_method}{How to handle requested dates outside of requested
dataset available data range. If "strict" (default),
an error will be returned. If "overlap", the requested dates will be
'truncated to the date range available in all requested datasets. If
"all", then the requested dates will be truncated to the available date
range per dataset. Non-temporal datasets are always returned.}

\item{ri_method}{The interpolation method used for extracting point
values. Available methods: "nearest" or "linear". Default is
"nearest".}

\item{dsn}{The detination directory to download and extract the return zipped
results. If NULL, the working directory is used.}

\item{req_name}{The request name. Used to name the folder containing the
downloaded data. If NULL, will be 'subset' plus a string of random
characters.}
}
\value{
A vector of paths to the downloaded data.
}
\description{
This function queries the GeoCDL REST API and returns
geospatial data matching the dataset variables and
spatio-temporal filters.
}
\details{
# Temporal subsets

For specifying the dates of your data request, you can either use `dates` for
simple date ranges or a combination of `years`, `months`, and `days` for
more complicated date patterns. For non-temporal datasets, these parameters
can be omitted and will be ignored if specified. To specify if you want
annual, monthly, or daily data, you specify your temporal subsets with only
year information (e.g. `dates=YYYY` or `years=YYYY`), or year and month
information (e.g. `dates=YYYY-MM` or `years=YYYY, months=MM`), or year,
(month), and day information (e.g. `dates=YYYY-MM-DD` or
`years=YYYY, months=MM, days=DD`). If not all datasets you request have the
same temporal grain, e.g. you request monthly data but one dataset only has
daily data or only has annual data, you can specify `grain_method` to choose
an alternative temporal grain. Use `validate_method` to specify how to handle
datasets whose temporal range does not match your requested temporal subset.
}
\seealso{
\code{\link{list_datasets}} to list available datasets and their IDs.
\code{\link{upload_geometry}} to obtain a geometry upload ID.
}
