% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List available GeoCDL datasets}
\usage{
list_datasets()
}
\value{
A data.frame of dataset names and IDs
}
\description{
This function queries the GeoCDL REST API and returns
currently supported dataset names and IDs.
}
\seealso{
\code{\link{view_metadata}} to list the dataset
  metadata by dataset ID
}
