% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_metadata.R
\name{view_metadata}
\alias{view_metadata}
\title{Request dataset metadata}
\usage{
view_metadata(ds)
}
\arguments{
\item{ds}{The GeoCDL dataset ID}
}
\value{
NULL, the metadata for the dataset as a printed message
}
\description{
This function queries the GeoCDL REST API and returns
the entire metadata associated with a dataset ID
}
\examples{
\dontrun{
view_metadata('PRISM')
}

}
\seealso{
\code{\link{list_datasets}} to list available datasets and their IDs
}
