% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_subset.R
\name{download_subset}
\alias{download_subset}
\title{Request dataset subset}
\usage{
download_subset(
  ds,
  vars,
  date_start = NULL,
  date_end = NULL,
  t_crs = NULL,
  t_geom = NULL
)
}
\arguments{
\item{ds}{The GeoCDL dataset ID.}

\item{vars}{A vector of variables from 'ds' to include, as available from
\code{\link{view_metadata}}.}

\item{date_start}{The start date of desired data. Use format 'YYYY'
for annual, 'YYYY-MM' for monthly, and 'YYYY-MM-DD' for daily data.
If NULL, returns entire temporal range.}

\item{date_end}{The end date of desired data. Use format 'YYYY'
for annual, 'YYYY-MM' for monthly, and 'YYYY-MM-DD' for daily data.
If NULL, returns entire temporal range.}

\item{t_crs}{The target CRS specified as an EPSG code. If not specified,
the CRS of the dataset will be used.}

\item{t_geom}{The area of interest specified as upper left and
lower right corners: c(c(xmin,ymax),(xmax,ymin)).These coordinates
are in `t_crs` if specified or the CRS of the first dataset.}
}
\value{
A vector of paths to the downloade data.
}
\description{
This function queries the GeoCDL REST API and returns
geospatial data matching the dataset variables and
spatio-temporal filters.
}
\examples{
\dontrun{
# Monthly PRISM precipitation from 2020 and in box of latlon coordinates
download_subset('PRISM','ppt','2020-01','2020-12',"4326",list(c(-109,37),c(-103,31)))
}

}
\seealso{
\code{\link{list_datasets}} to list available datasets and their IDs.
}
