% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_geometry.R
\name{upload_geometry}
\alias{upload_geometry}
\title{Upload a user geometry}
\usage{
upload_geometry(geom)
}
\arguments{
\item{geom}{A sp or sf object or a filename. ELABORATE}
}
\value{
A geometry upload ID string
}
\description{
This function uploads a user geometry to the GeoCDL
REST API and returns a geometry upload ID to use
in subset requests.
}
\seealso{
\code{\link{view_metadata}} to list the dataset
  metadata by dataset ID
}
